supabase: refine severity precision and lock down anon write paths
- 0001_init.sql: pin severity column changed from integer to numeric(3,1).
  Gemini returns scores rounded to 0.1, so 7.8 and 7.0 should be
  distinguishable for ranking and sub-integer thresholds.
- 0002_rpc_acknowledge_resolve.sql: drop the permissive
  anon_update_status RLS policy. It only constrained the final status
  value, leaving anon free to modify any other column. Replace with
  acknowledge_pin/resolve_pin SECURITY DEFINER functions and revoke
  table-level UPDATE on pins from anon.
- 0003_revoke_rpc_from_anon.sql: revoke EXECUTE on the new RPCs from
  anon and authenticated, since the convention is now to route pin
  mutations through FastAPI under service_role rather than directly
  from the browser.
- 0004_drop_rpc_and_lock_default_privileges.sql: drop the RPCs entirely
  after FastAPI endpoints landed. Postgres CREATE FUNCTION implicitly
  grants EXECUTE to PUBLIC and anon inherits PUBLIC, which kept the
  RPCs callable from anon despite the prior REVOKE. ALTER DEFAULT
  PRIVILEGES IN SCHEMA public REVOKE EXECUTE ON FUNCTIONS FROM PUBLIC
  prevents future functions from reopening the same hole.

+++ supabase/migrations/0001_init.sql
@@ -28,7 +28,7 @@ CREATE TABLE public.pins (
   lon             double precision NOT NULL,
   milepost        text,
   defect_type     text NOT NULL,
-  severity        integer NOT NULL,
+  severity        numeric(3,1) NOT NULL,
   confidence      real NOT NULL,
   device_id       text NOT NULL,
   frame_id        bigint,
